+++ src/groupsmembers/groupsmembers-service.js
@@ -16,6 +16,7 @@ const GroupsMembersService = {
         'gm.member_id',
         'u.username',
         'u.fullname',
+        'u.email',
         'gm.group_id',
         'g.name',
       )
