+++ src/tasks/tasks-router.js
@@ -2,6 +2,10 @@ const express = require('express');
 const path = require('path');
 const xss = require('xss');
 const TasksService = require('./tasks-service.js');
+const GroupsMembersService = require('../groupsmembers/groupsmembers-service');
+
+const nodemailer = require('nodemailer');
+const { transporter } = require('../mail-service');
 
 const tasksRouter = express.Router();
 const jsonParser = express.json();
@@ -69,8 +73,8 @@ tasksRouter.post('/', requireAuth, jsonParser, async (req, res, next) => {
     date_due,
     user_assigned_id,
     group_id,
-    category_id,
-    priority,
+    // category_id,
+    // priority,
   } = req.body;
 
   for (const field of [
@@ -79,7 +83,7 @@ tasksRouter.post('/', requireAuth, jsonParser, async (req, res, next) => {
     'date_due',
     'user_assigned_id',
     'group_id',
-    'priority',
+    // 'priority',
   ])
     if (!req.body[field])
       return res.status(400).json({
@@ -94,8 +98,8 @@ tasksRouter.post('/', requireAuth, jsonParser, async (req, res, next) => {
     date_due,
     user_assigned_id,
     group_id,
-    category_id,
-    priority,
+    // category_id,
+    // priority,
   };
   try {
     const newTask = await TasksService.postNewTask(
@@ -161,7 +165,40 @@ tasksRouter
         task_id,
         updateInfo,
       );
-      res.status(200).json(taskFormat(updatedTask));
+      if (updatedTask) {
+        const group_id = updatedTask.group_id;
+
+        // get all member emails in tthe group
+        // that have email notifcations on (do later)
+        const groupUsers = await GroupsMembersService.getGroupMembers(
+          req.app.get('db'),
+          group_id,
+        );
+        const emails = groupUsers.map(user => user.email);
+        let allMailOptions = emails.map(email => {
+          return {
+            from: '"13 Minutes" <groopnotify@gmail.com>',
+            to: email,
+            subject: `Task ${updatedTask.name} has been updated`,
+            html: `<section style="margin: 0 auto;"><div style="max-width: 600px; margin: 0 auto; padding: 2rem; text-align: center; background-color: #363432; color: #fafafa; "><h2>Groop</h2><div style="height: 0; width: 200px; border: 1px solid #4a9afa;"></div><h1>The following task has been updated</h1><div style="text-align: left;"><p>${updatedTask.name}</p><p>${updatedTask.description}</p><p>completed: ${updatedTask.completed}</p></div></div></section>`,
+          };
+        });
+
+        allMailOptions.forEach(async mailOption => {
+          return (info = await transporter.sendMail(mailOption, function(
+            error,
+            info,
+          ) {
+            if (error) return false;
+            else {
+              console.log('Message sent: ' + info.response);
+              return true;
+            }
+          }));
+        });
+
+        res.status(200).json(taskFormat(updatedTask));
+      }
     } catch (error) {
       next(error);
     }
